+++ mobiledrivinglicence/src/commonMain/kotlin/at/asitplus/wallet/mdl/Initializer.kt
@@ -45,7 +45,7 @@ object Initializer {
                 MobileDrivingLicenceDataElements.AGE_IN_YEARS to UInt.serializer(),
                 MobileDrivingLicenceDataElements.AGE_BIRTH_YEAR to UInt.serializer(),
                 MobileDrivingLicenceDataElements.SIGNATURE_USUAL_MARK to ByteArraySerializer(),
-                )
+            )
         )
     }
